+++ packages/domain-social/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/contracts@4.0.4
+
 ## 0.3.3
 
 ### Patch Changes